requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
diffusers==0.26.3
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
pydantic==1.10.13
safetensors==0.4.1
huggingface-hub==0.20.2
numpy<2.0